bind Prometheus exporter to new port
By default, the Prometheus exporter will add a route for `/metrics` to
the running app. In our case, the use of SERVER_NAME results in 404s for
these requests. This is an attempt to workaround that. If I understand
correctly, this change provides us with a new app context which avoids
that problem.

In testing, I can access the `/metrics` endpoint with any host header
and the server responds with the metrics.