Remove app name truncation
This changes the SSO Dashboard behavior to stop truncating app names.
Instead this
* leaves the full app name intact (for accessibility)
* achieves the visual truncation through CSS
* adds a tooltip to the name with the full value

This will improve accessibility and allow users to see the full name of the app without affecting the grid layout. Without this change, it's impossible for users to determine the full name of an app that is over 18 characters long (by any means). The data just isn't there.