Update apps.yml
This apps.yml is not used as the real apps.yml, located at https://github.com/mozilla-iam/sso-dashboard-configuration/blob/master/apps.yml is fetched from S3.

We should probably get rid of the `apps.yml` that's in the sso-dashboard repo as it just causes confusion (as it's not actually used).

For the moment I'm just going to update to the head of master in sso-dashboard-configuration ( https://github.com/mozilla-iam/sso-dashboard-configuration/commit/598bba52f6cdc2bce9f96281e663e1609dec2eb7 )