Support RP-Initiated logout, further removing the need for NLX
Flask-pyoidc supports provider discovery which we can make use of for
auto-enabling the RP-Initiated Logout End Session Endpoint Discovery feature
set.

When enabled the logout flow changes a bit and does not require a redirect in
NLX, insted depending directly on Auth0 to complete the OIDC ceremony.

Prerequisites:

* `/logout` URL added as an accepted logout URL in the _client_ settings;
* RP-Initiated Logout End Session Endpoint Discovery enabled in _tenant_ settings;
    * Toggling this in Auth0 _will require that you restart the application_.

Jira: IAM-1493