models Add a schema for application definitions
There was some discussion about schemas on GitHub and Slack. This commit takes
some of those thoughts and codifies them.

The types defined in `dashboard.models.apps` are copied to the configuration
repository, where they're used in the tests.

A Better :tm: solution would maybe include publishing a package so we don't
need to depend on copying/pasting the schemas, but for now this'll do.

Jira: IAM-950