Environment variables were not shell friendly
The environment names were dictated by some library we no longer use. I removed
that dependency, but kept the environment variables the same to not break stuff
anything.

This commit changes that to be shell-friendly. Before, if you wanted to write a
script to run the app, you'd need to resort to using either an external app, or
`env -S $(< envfile) flask ...`.

By swapping to dashes we're a little more generic in the kinds of ways you can
call the app.

Jira: IAM-950